+++ packages/rsocket-router/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @powersync/service-rsocket-router
 
+## 0.0.10
+
+### Patch Changes
+
+- 9bff878: Fix issue where WebSocket close events would not immediately propagate to router handlers.
+
+## 0.0.9
+
+### Patch Changes
+
+- 909f71a: Fix concurrent connection limiting for websockets
+- Updated dependencies [909f71a]
+  - @powersync/lib-services-framework@0.1.1
+
 ## 0.0.8
 
 ### Patch Changes