+++ service/CHANGELOG.md
@@ -1,4 +1,27 @@
-# powersync-open-service
+# @powersync/service-image
+
+## 0.3.2
+
+### Patch Changes
+
+- bdbf95c: Log user_id and sync stats for each connection
+- Updated dependencies [876f4a0]
+- Updated dependencies [9bff878]
+- Updated dependencies [bdbf95c]
+  - @powersync/service-sync-rules@0.18.1
+  - @powersync/service-rsocket-router@0.0.10
+  - @powersync/service-core@0.4.2
+
+## 0.3.1
+
+### Patch Changes
+
+- 909f71a: Fix concurrent connection limiting for websockets
+- Updated dependencies [909f71a]
+- Updated dependencies [1066f86]
+  - @powersync/service-rsocket-router@0.0.9
+  - @powersync/lib-services-framework@0.1.1
+  - @powersync/service-core@0.4.1
 
 ## 0.3.0